+++ doc/CHANGELOG.rst
@@ -4,9 +4,13 @@ Changelog
 current
 -------
 
+- Add Authentication using Flask-Login and rauth
+    - Google OAuth2
+    - Facebook OAuth2
+
 v0.1
 ----
 
 - Initial release
-- Parking slots for Montréal
-- Parking slots for Québec
+- Parking free slots for Montréal
+- Parking free slots for Québec